AudienceConfig moved to config package
